// frontend/src/shared/lib/api/model/siblingRenameResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Cellar
 * OpenAPI spec version: 1.3.0-17-gd9da37bb-dirty
 */
import type { SiblingRenameResponseCode } from './siblingRenameResponseCode';
import type { SiblingRenameResponseName } from './siblingRenameResponseName';
import type { SiblingRenameResponseError } from './siblingRenameResponseError';

export interface SiblingRenameResponse {
  protocol_id: string;
  code: SiblingRenameResponseCode;
  name: SiblingRenameResponseName;
  error: SiblingRenameResponseError;
}